shift more responsibility to the backend
