+++ src/checkout/hooks/useCreatePaymentSession.ts
@@ -1,14 +1,14 @@
-import { createSession } from "checkout/lib";
+import { createPaymentSession } from "checkout/lib";
 import { useCallback, useState } from "react";
 import { useAdvertiser } from "auth/hooks/queries/useAdvertiser";
 
-export function useCreateSession() {
+export function useCreatePaymentSession() {
   const [loading, setLoading] = useState(false);
 
   const { advertiser } = useAdvertiser();
   const replaceSession = useCallback(async (campaignId: string) => {
     setLoading(true);
-    await createSession(advertiser.id, campaignId ?? "")
+    await createPaymentSession(advertiser.id, campaignId ?? "")
       .then((url) => {
         window.location.replace(url);
       })
@@ -18,5 +18,5 @@ export function useCreateSession() {
       });
   }, []);
 
-  return { replaceSession, loading };
+  return { createPaymentSession: replaceSession, loading };
 }
